requirements.txt
pillow>=9.5.0
protobuf>=4.23.2
addict
future
lmdb
numpy
opencv-python
Pillow
pyyaml
requests
scikit-image
scipy
tb-nightly
torch>=1.7.1
torchvision
tqdm
yapf
lpips
insightface>=0.7.3
